Refactor to remove duplication of journey helper methods
