+++ app/controllers/base_public_controller.rb
@@ -2,17 +2,14 @@ class BasePublicController < ApplicationController
   include DfE::Analytics::Requests
   include ClaimSessionTimeout
   include HttpAuthConcern
+  include JourneyConcern
 
-  helper_method :current_journey_routing_name, :claim_timeout_in_minutes
+  helper_method :claim_timeout_in_minutes
   before_action :add_view_paths
   before_action :end_expired_claim_sessions
   after_action :update_last_seen_at
 
   private
-
-  def current_journey_routing_name
-    params[:journey]
-  end
 
   def end_expired_claim_sessions
     if claim_session_timed_out?
@@ -26,7 +23,6 @@ class BasePublicController < ApplicationController
   end
 
   def add_view_paths
-    journey = Journeys.for_routing_name(current_journey_routing_name)
-    prepend_view_path(Rails.root.join("app", "views", journey::VIEW_PATH)) if journey
+    prepend_view_path(Rails.root.join("app", "views", journey::VIEW_PATH))
   end
 end
