+++ app/controllers/omniauth_callbacks_controller.rb
@@ -1,14 +1,12 @@
 class OmniauthCallbacksController < ApplicationController
-  include PartOfClaimJourney
-  skip_before_action :check_whether_closed_for_submissions
-  skip_before_action :send_unstarted_claimants_to_the_start
+  include JourneyConcern
 
   def callback
     auth = request.env["omniauth.auth"]
 
     session[:user_info] = auth.extra.raw_info
 
-    redirect_to claim_path(journey: Journeys.for_policy(policy)::ROUTING_NAME, slug: "teacher-detail")
+    redirect_to claim_path(journey: current_journey_routing_name, slug: "teacher-detail")
   end
 
   def failure
