+++ app/controllers/static_pages_controller.rb
@@ -12,7 +12,6 @@ class StaticPagesController < BasePublicController
   end
 
   def landing_page
-    journey = Journeys.for_routing_name(current_journey_routing_name)
     @academic_year = journey.configuration.current_academic_year
     @journey_open = journey.configuration.open_for_submissions?
